+++ src/curies/__init__.py
@@ -6,6 +6,7 @@ from .api import (
     DuplicateURIPrefixes,
     DuplicateValueError,
     NamedReference,
+    Prefix,
     Record,
     Records,
     Reference,
@@ -30,7 +31,7 @@ from .sources import (
     get_obo_converter,
     get_prefixcommons_converter,
 )
-from .typr import CURIE, URI, Prefix
+from .typr import CURIE, URI
 from .version import get_version
 
 __all__ = [
