+++ src/curies/api.py
@@ -1,5 +1,7 @@
 """Data structures and algorithms for :mod:`curies`."""
 
+from __future__ import annotations
+
 import csv
 import itertools as itt
 import json
@@ -15,15 +17,16 @@ from typing import (
     Callable,
     Literal,
     NamedTuple,
-    Optional,
     TypeVar,
     Union,
     cast,
     overload,
 )
 
-from pydantic import BaseModel, ConfigDict, Field, RootModel, field_validator
+from pydantic import BaseModel, ConfigDict, Field, GetCoreSchemaHandler, RootModel, field_validator
+from pydantic_core import core_schema
 from pytrie import StringTrie
+from typing_extensions import Self
 
 if TYPE_CHECKING:  # pragma: no cover
     import pandas
@@ -35,6 +38,7 @@ __all__ = [
     "DuplicateURIPrefixes",
     "DuplicateValueError",
     "NamedReference",
+    "Prefix",
     "Record",
     "Records",
     "Reference",
@@ -55,6 +59,10 @@ logger = logging.getLogger(__name__)
 
 X = TypeVar("X")
 LocationOr = Union[str, Path, X]
+
+#: From https://stackoverflow.com/a/55038380, with the addition
+#: of being able to match an empty string
+NCNAME_RE = r"^$|^[a-zA-Z_][\w.-]*$"
 
 
 def _get_field_validator_values(values, key: str):  # type:ignore
@@ -133,7 +141,7 @@ class ReferenceTuple(NamedTuple):
         return f"{self.prefix}:{self.identifier}"
 
     @classmethod
-    def from_curie(cls, curie: str, *, sep: str = ":") -> "ReferenceTuple":
+    def from_curie(cls, curie: str, *, sep: str = ":") -> ReferenceTuple:
         """Parse a CURIE string and populate a reference tuple.
 
         :param curie: A string representation of a compact URI (CURIE)
@@ -145,6 +153,125 @@ class ReferenceTuple(NamedTuple):
         """
         prefix, identifier = _split(curie, sep=sep)
         return cls(prefix, identifier)
+
+
+class Prefix(str):
+    """A string that is validated by Pydantic as a CURIE prefix.
+
+    This class is a subclass of Python's built-in string class,
+    so you can wrap any string with it:
+
+    .. code-block:: python
+
+        from curies import Prefix
+
+        prefix = Prefix("CHEBI")
+
+    You can implicitly type annotate data with this class:
+
+    .. code-block:: python
+
+        from curies import Prefix
+
+        prefix_map: dict[Prefix, str] = {
+            "CHEBI": "http://purl.obolibrary.org/obo/CHEBI_",
+        }
+
+    When used inside a Pydantic model, this class knows how to
+    do validation that the prefix matches the regular expression
+    for an XSD NCName. Here's an example usage with Pydantic:
+
+    .. code-block:: python
+
+        from curies import Prefix
+        from pydantic import BaseModel
+
+
+        class ResourceInfo(BaseModel):
+            prefix: Prefix
+            name: str
+
+
+        model = ResourceInfo.model_validate(
+            {
+                "prefix": "CHEBI",
+                "name": "Chemical Entities of Biological Interest",
+            }
+        )
+
+        # raises a pydantic.ValidationError, because the prefix
+        # doesn't match the NCName pattern
+        ResourceInfo.model_validate(
+            {
+                "prefix": "$nope",
+                "name": "An invalid semantic space!",
+            }
+        )
+
+    This class implements a hook that uses Pydantic's "context"
+    for validation that lets you pass a :class:`Converter` to check
+    for existence and standardization with respect to the context
+    in the converter:
+
+    .. code-block:: python
+
+        from curies import Prefix, get_obo_converter
+        from pydantic import BaseModel
+
+        class ResourceInfo(BaseModel):
+            prefix: Prefix
+            name: str
+
+        converter = get_obo_converter()
+        model = ResourceInfo.model_validate(
+            {
+                "prefix": "CHEBI",
+                "name": "Chemical Entities of Biological Interest",
+            },
+            context=converter,
+        )
+
+        # raises a pydantic.ValidationError, because the prefix
+        # is not registered in the OBO Foundry, and is therefore
+        # not part of the OBO converter
+        ResourceInfo.model_validate(
+            {
+                "prefix": "efo",
+                "name": "Experimental Factor Ontology",
+            },
+            context=converter,
+        )
+
+        # In case you need to pass more arbitrary
+        # context, you can also use a dict with the key
+        # "converter"
+        ResourceInfo.model_validate(
+            {
+                "prefix": "CHEBI",
+                "name": "Chemical Entities of Biological Interest",
+            },
+            context={
+                "converter": converter,
+                ...
+            },
+        )
+    """
+
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls, source: type[Any], handler: GetCoreSchemaHandler
+    ) -> core_schema.AfterValidatorFunctionSchema:
+        return core_schema.with_info_after_validator_function(
+            cls._validate,
+            core_schema.str_schema(pattern=NCNAME_RE, strict=False),
+        )
+
+    @classmethod
+    def _validate(cls, __input_value: str, info: core_schema.ValidationInfo) -> Self:
+        converter = _get_converter_from_context(info)
+        if converter is None:
+            return cls(__input_value)
+        return cls(converter.standardize_prefix(__input_value, strict=True))
 
 
 class Reference(BaseModel):
@@ -191,7 +318,7 @@ class Reference(BaseModel):
     ReferenceTuple(prefix='chebi', identifier='1234')
     """
 
-    prefix: str = Field(
+    prefix: Prefix = Field(
         ...,
         description="The prefix used in a compact URI (CURIE).",
     )
@@ -201,7 +328,7 @@ class Reference(BaseModel):
 
     model_config = ConfigDict(frozen=True)
 
-    def __lt__(self, other: "Reference") -> bool:
+    def __lt__(self, other: Reference) -> bool:
         """Sort the reference lexically first by prefix, then by identifier."""
         return self.pair < other.pair
 
@@ -233,18 +360,19 @@ class Reference(BaseModel):
         return ReferenceTuple(self.prefix, self.identifier)
 
     @classmethod
-    def from_curie(cls, curie: str, *, sep: str = ":") -> "Reference":
+    def from_curie(cls, curie: str, *, sep: str = ":", converter: Converter | None = None) -> Reference:
         """Parse a CURIE string and populate a reference.
 
         :param curie: A string representation of a compact URI (CURIE)
         :param sep: The separator
+        :param converter: The converter to use as context when parsing
         :return: A reference object
 
         >>> Reference.from_curie("chebi:1234")
         Reference(prefix='chebi', identifier='1234')
         """
         prefix, identifier = _split(curie, sep=sep)
-        return cls(prefix=prefix, identifier=identifier)
+        return cls.model_validate({"prefix": prefix, "identifier": identifier}, context=converter)
 
 
 class NamedReference(Reference):
@@ -257,7 +385,7 @@ class NamedReference(Reference):
     model_config = ConfigDict(frozen=True)
 
     @classmethod
-    def from_curie(cls, curie: str, name: str, *, sep: str = ":") -> "NamedReference":  # type:ignore
+    def from_curie(cls, curie: str, name: str, *, sep: str = ":") -> NamedReference:  # type:ignore
         """Parse a CURIE string and populate a reference.
 
         :param curie: A string representation of a compact URI (CURIE)
@@ -293,10 +421,10 @@ class Record(BaseModel):
     )
     prefix_synonyms: list[str] = Field(default_factory=list, title="CURIE prefix synonyms")
     uri_prefix_synonyms: list[str] = Field(default_factory=list, title="URI prefix synonyms")
-    pattern: Optional[str] = Field(
+    pattern: str | None = Field(
         default=None,
         description="The regular expression pattern for entries in this semantic space. "
-        "Warning: this is an experimental feature.",
+                    "Warning: this is an experimental feature.",
     )
 
     @field_validator("prefix_synonyms")  # type:ignore
@@ -647,8 +775,8 @@ class Converter:
         self,
         prefix: str,
         uri_prefix: str,
-        prefix_synonyms: Optional[Collection[str]] = None,
-        uri_prefix_synonyms: Optional[Collection[str]] = None,
+        prefix_synonyms: Collection[str] | None = None,
+        uri_prefix_synonyms: Collection[str] | None = None,
         *,
         case_sensitive: bool = True,
         merge: bool = False,
@@ -700,8 +828,8 @@ class Converter:
 
     @classmethod
     def from_extended_prefix_map(
-        cls, records: LocationOr[Iterable[Union[Record, dict[str, Any]]]], **kwargs: Any
-    ) -> "Converter":
+        cls, records: LocationOr[Iterable[Record | dict[str, Any]]], **kwargs: Any
+    ) -> Converter:
         """Get a converter from a list of dictionaries by creating records out of them.
 
         :param records:
@@ -778,7 +906,7 @@ class Converter:
     @classmethod
     def from_priority_prefix_map(
         cls, data: LocationOr[Mapping[str, list[str]]], **kwargs: Any
-    ) -> "Converter":
+    ) -> Converter:
         """Get a converter from a priority prefix map.
 
         :param data:
@@ -815,9 +943,7 @@ class Converter:
         )
 
     @classmethod
-    def from_prefix_map(
-        cls, prefix_map: LocationOr[Mapping[str, str]], **kwargs: Any
-    ) -> "Converter":
+    def from_prefix_map(cls, prefix_map: LocationOr[Mapping[str, str]], **kwargs: Any) -> Converter:
         """Get a converter from a simple prefix map.
 
         :param prefix_map:
@@ -855,7 +981,7 @@ class Converter:
     @classmethod
     def from_reverse_prefix_map(
         cls, reverse_prefix_map: LocationOr[Mapping[str, str]], **kwargs: Any
-    ) -> "Converter":
+    ) -> Converter:
         """Get a converter from a reverse prefix map.
 
         :param reverse_prefix_map:
@@ -900,7 +1026,7 @@ class Converter:
         return cls(records, **kwargs)
 
     @classmethod
-    def from_jsonld(cls, data: LocationOr[dict[str, Any]], **kwargs: Any) -> "Converter":
+    def from_jsonld(cls, data: LocationOr[dict[str, Any]], **kwargs: Any) -> Converter:
         """Get a converter from a JSON-LD object, which contains a prefix map in its ``@context`` key.
 
         :param data:
@@ -938,7 +1064,7 @@ class Converter:
     @classmethod
     def from_jsonld_github(
         cls, owner: str, repo: str, *path: str, branch: str = "main", **kwargs: Any
-    ) -> "Converter":
+    ) -> Converter:
         """Construct a remote JSON-LD URL on GitHub then parse with :meth:`Converter.from_jsonld`.
 
         :param owner: A github repository owner or organization (e.g., ``biopragmatics``)
@@ -971,9 +1097,9 @@ class Converter:
     @classmethod
     def from_rdflib(
         cls,
-        graph_or_manager: Union["rdflib.Graph", "rdflib.namespace.NamespaceManager"],
+        graph_or_manager: rdflib.Graph | rdflib.namespace.NamespaceManager,
         **kwargs: Any,
-    ) -> "Converter":
+    ) -> Converter:
         """Get a converter from an RDFLib graph or namespace manager.
 
         :param graph_or_manager: A RDFLib graph or manager object
@@ -1004,10 +1130,10 @@ class Converter:
     @classmethod
     def from_shacl(
         cls,
-        graph: Union[str, Path, "rdflib.Graph"],
-        format: Optional[str] = None,
+        graph: str | Path | rdflib.Graph,
+        format: str | None = None,
         **kwargs: Any,
-    ) -> "Converter":
+    ) -> Converter:
         """Get a converter from SHACL, either in a turtle f.
 
         :param graph: A RDFLib graph, a Path, a string representing a file path, or a string URL
@@ -1107,7 +1233,8 @@ class Converter:
     @overload
     def compress_or_standardize(
         self, uri_or_curie: str, *, strict: Literal[True] = True, passthrough: bool = ...
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
@@ -1117,7 +1244,8 @@ class Converter:
         *,
         strict: Literal[False] = False,
         passthrough: Literal[True] = True,
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
@@ -1127,11 +1255,12 @@ class Converter:
         *,
         strict: Literal[False] = False,
         passthrough: Literal[False] = False,
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
     def compress_or_standardize(
         self, uri_or_curie: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    ) -> str | None:
         """Compress a URI or standardize a CURIE.
 
         :param uri_or_curie:
@@ -1188,23 +1317,24 @@ class Converter:
     @overload
     def compress(
         self, uri: str, *, strict: Literal[True] = True, passthrough: bool = ...
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def compress(
         self, uri: str, *, strict: Literal[False] = False, passthrough: Literal[True] = True
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def compress(
         self, uri: str, *, strict: Literal[False] = False, passthrough: Literal[False] = False
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
-    def compress(
-        self, uri: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    def compress(self, uri: str, *, strict: bool = False, passthrough: bool = False) -> str | None:
         """Compress a URI to a CURIE, if possible.
 
         :param uri:
@@ -1250,7 +1380,7 @@ class Converter:
             return uri
         return None
 
-    def parse_uri(self, uri: str) -> Union[ReferenceTuple, tuple[None, None]]:
+    def parse_uri(self, uri: str) -> ReferenceTuple | tuple[None, None]:
         """Compress a URI to a CURIE pair.
 
         :param uri:
@@ -1276,7 +1406,7 @@ class Converter:
         except KeyError:
             return None, None
         else:
-            return ReferenceTuple(prefix, uri[len(value) :])
+            return ReferenceTuple(prefix, uri[len(value):])
 
     def is_curie(self, s: str) -> bool:
         """Check if the string can be parsed as a CURIE by this converter.
@@ -1310,7 +1440,8 @@ class Converter:
     @overload
     def expand_or_standardize(
         self, curie_or_uri: str, *, strict: Literal[True] = True, passthrough: bool = ...
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
@@ -1320,7 +1451,8 @@ class Converter:
         *,
         strict: Literal[False] = False,
         passthrough: Literal[True] = True,
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
@@ -1330,11 +1462,12 @@ class Converter:
         *,
         strict: Literal[False] = False,
         passthrough: Literal[False] = False,
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
     def expand_or_standardize(
         self, curie_or_uri: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    ) -> str | None:
         """Expand a CURIE or standardize a URI.
 
         :param curie_or_uri:
@@ -1391,23 +1524,24 @@ class Converter:
     @overload
     def expand(
         self, curie: str, *, strict: Literal[True] = True, passthrough: bool = ...
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def expand(
         self, curie: str, *, strict: Literal[False] = False, passthrough: Literal[True] = True
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def expand(
         self, curie: str, *, strict: Literal[False] = False, passthrough: Literal[False] = False
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
-    def expand(
-        self, curie: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    def expand(self, curie: str, *, strict: bool = False, passthrough: bool = False) -> str | None:
         """Expand a CURIE to a URI, if possible.
 
         :param curie:
@@ -1443,7 +1577,7 @@ class Converter:
             return curie
         return None
 
-    def expand_all(self, curie: str) -> Optional[Collection[str]]:
+    def expand_all(self, curie: str) -> Collection[str] | None:
         """Expand a CURIE pair to all possible URIs.
 
         :param curie:
@@ -1473,7 +1607,7 @@ class Converter:
         """Parse a CURIE."""
         return ReferenceTuple.from_curie(curie, sep=self.delimiter)
 
-    def expand_pair(self, prefix: str, identifier: str) -> Optional[str]:
+    def expand_pair(self, prefix: str, identifier: str) -> str | None:
         """Expand a CURIE pair to the standard URI.
 
         :param prefix:
@@ -1500,7 +1634,7 @@ class Converter:
             return None
         return uri_prefix + identifier
 
-    def expand_pair_all(self, prefix: str, identifier: str) -> Optional[Collection[str]]:
+    def expand_pair_all(self, prefix: str, identifier: str) -> Collection[str] | None:
         """Expand a CURIE pair to all possible URIs.
 
         :param prefix:
@@ -1537,23 +1671,26 @@ class Converter:
     @overload
     def standardize_prefix(
         self, prefix: str, *, strict: Literal[True] = True, passthrough: bool = False
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def standardize_prefix(
         self, prefix: str, *, strict: Literal[False] = False, passthrough: Literal[True] = True
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def standardize_prefix(
         self, prefix: str, *, strict: Literal[False] = False, passthrough: Literal[False] = False
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
     def standardize_prefix(
         self, prefix: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    ) -> str | None:
         """Standardize a prefix.
 
         :param prefix:
@@ -1595,23 +1732,26 @@ class Converter:
     @overload
     def standardize_curie(
         self, curie: str, *, strict: Literal[True] = True, passthrough: bool = False
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def standardize_curie(
         self, curie: str, *, strict: Literal[False] = False, passthrough: Literal[True] = True
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def standardize_curie(
         self, curie: str, *, strict: Literal[False] = False, passthrough: Literal[False] = False
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
     def standardize_curie(
         self, curie: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    ) -> str | None:
         """Standardize a CURIE.
 
         :param curie:
@@ -1660,23 +1800,26 @@ class Converter:
     @overload
     def standardize_uri(
         self, uri: str, *, strict: Literal[True] = True, passthrough: bool = False
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def standardize_uri(
         self, uri: str, *, strict: Literal[False] = False, passthrough: Literal[True] = True
-    ) -> str: ...
+    ) -> str:
+        ...
 
     # docstr-coverage:excused `overload`
     @overload
     def standardize_uri(
         self, uri: str, *, strict: Literal[False] = False, passthrough: Literal[False] = False
-    ) -> Optional[str]: ...
+    ) -> str | None:
+        ...
 
     def standardize_uri(
         self, uri: str, *, strict: bool = False, passthrough: bool = False
-    ) -> Optional[str]:
+    ) -> str | None:
         """Standardize a URI.
 
         :param uri:
@@ -1727,9 +1870,9 @@ class Converter:
 
     def pd_compress(
         self,
-        df: "pandas.DataFrame",
-        column: Union[str, int],
-        target_column: Union[None, str, int] = None,
+        df: pandas.DataFrame,
+        column: str | int,
+        target_column: None | str | int = None,
         strict: bool = False,
         passthrough: bool = False,
         ambiguous: bool = False,
@@ -1750,9 +1893,9 @@ class Converter:
 
     def pd_expand(
         self,
-        df: "pandas.DataFrame",
-        column: Union[str, int],
-        target_column: Union[None, str, int] = None,
+        df: pandas.DataFrame,
+        column: str | int,
+        target_column: None | str | int = None,
         strict: bool = False,
         passthrough: bool = False,
         ambiguous: bool = False,
@@ -1773,10 +1916,10 @@ class Converter:
 
     def pd_standardize_prefix(
         self,
-        df: "pandas.DataFrame",
+        df: pandas.DataFrame,
         *,
-        column: Union[str, int],
-        target_column: Union[None, str, int] = None,
+        column: str | int,
+        target_column: None | str | int = None,
         strict: bool = False,
         passthrough: bool = False,
     ) -> None:
@@ -1794,10 +1937,10 @@ class Converter:
 
     def pd_standardize_curie(
         self,
-        df: "pandas.DataFrame",
+        df: pandas.DataFrame,
         *,
-        column: Union[str, int],
-        target_column: Union[None, str, int] = None,
+        column: str | int,
+        target_column: None | str | int = None,
         strict: bool = False,
         passthrough: bool = False,
     ) -> None:
@@ -1829,10 +1972,10 @@ class Converter:
 
     def pd_standardize_uri(
         self,
-        df: "pandas.DataFrame",
+        df: pandas.DataFrame,
         *,
-        column: Union[str, int],
-        target_column: Union[None, str, int] = None,
+        column: str | int,
+        target_column: None | str | int = None,
         strict: bool = False,
         passthrough: bool = False,
     ) -> None:
@@ -1850,10 +1993,10 @@ class Converter:
 
     def file_compress(
         self,
-        path: Union[str, Path],
+        path: str | Path,
         column: int,
         *,
-        sep: Optional[str] = None,
+        sep: str | None = None,
         header: bool = True,
         strict: bool = False,
         passthrough: bool = False,
@@ -1876,10 +2019,10 @@ class Converter:
 
     def file_expand(
         self,
-        path: Union[str, Path],
+        path: str | Path,
         column: int,
         *,
-        sep: Optional[str] = None,
+        sep: str | None = None,
         header: bool = True,
         strict: bool = False,
         passthrough: bool = False,
@@ -1902,10 +2045,10 @@ class Converter:
 
     @staticmethod
     def _file_helper(
-        func: Callable[[str], Optional[str]],
-        path: Union[str, Path],
+        func: Callable[[str], str | None],
+        path: str | Path,
         column: int,
-        sep: Optional[str] = None,
+        sep: str | None = None,
         header: bool = True,
     ) -> None:
         path = Path(path).expanduser().resolve()
@@ -1923,7 +2066,7 @@ class Converter:
                 writer.writerow(_header)
             writer.writerows(rows)
 
-    def get_record(self, prefix: str) -> Optional[Record]:
+    def get_record(self, prefix: str) -> Record | None:
         """Get the record for the prefix."""
         # TODO better data structure for this
         for record in self.records:
@@ -1931,7 +2074,7 @@ class Converter:
                 return record
         return None
 
-    def get_subconverter(self, prefixes: Iterable[str]) -> "Converter":
+    def get_subconverter(self, prefixes: Iterable[str]) -> Converter:
         r"""Get a converter with a subset of prefixes.
 
         :param prefixes: A list of prefixes to keep from this converter. These can
@@ -2093,7 +2236,7 @@ def load_prefix_map(prefix_map: LocationOr[Mapping[str, str]], **kwargs: Any) ->
 
 
 def load_extended_prefix_map(
-    records: LocationOr[Iterable[Union[Record, dict[str, Any]]]], **kwargs: Any
+    records: LocationOr[Iterable[Record | dict[str, Any]]], **kwargs: Any
 ) -> Converter:
     """Get a converter from a list of dictionaries by creating records out of them.
 
@@ -2181,7 +2324,7 @@ def load_jsonld_context(data: LocationOr[dict[str, Any]], **kwargs: Any) -> Conv
     return Converter.from_jsonld(data, **kwargs)
 
 
-def load_shacl(data: LocationOr["rdflib.Graph"], **kwargs: Any) -> Converter:
+def load_shacl(data: LocationOr[rdflib.Graph], **kwargs: Any) -> Converter:
     """Get a converter from a JSON-LD object, which contains a prefix map in its ``@context`` key.
 
     :param data:
@@ -2193,7 +2336,7 @@ def load_shacl(data: LocationOr["rdflib.Graph"], **kwargs: Any) -> Converter:
     return Converter.from_shacl(data, **kwargs)
 
 
-def write_extended_prefix_map(converter: Converter, path: Union[str, Path]) -> None:
+def write_extended_prefix_map(converter: Converter, path: str | Path) -> None:
     """Write an extended prefix map as JSON to a file."""
     path = _ensure_path(path)
     path.write_text(
@@ -2206,9 +2349,9 @@ def write_extended_prefix_map(converter: Converter, path: Union[str, Path]) -> N
     )
 
 
-def _record_to_dict(record: Record) -> Mapping[str, Union[str, list[str]]]:
+def _record_to_dict(record: Record) -> Mapping[str, str | list[str]]:
     """Convert a record to a dict."""
-    rv: dict[str, Union[str, list[str]]] = {
+    rv: dict[str, str | list[str]] = {
         "prefix": record.prefix,
         "uri_prefix": record.uri_prefix,
     }
@@ -2221,7 +2364,7 @@ def _record_to_dict(record: Record) -> Mapping[str, Union[str, list[str]]]:
     return rv
 
 
-def _ensure_path(path: Union[str, Path]) -> Path:
+def _ensure_path(path: str | Path) -> Path:
     if isinstance(path, str):
         path = Path(path).resolve()
     return path
@@ -2243,7 +2386,7 @@ def _get_jsonld_context(
 
 def write_jsonld_context(
     converter: Converter,
-    path: Union[str, Path],
+    path: str | Path,
     *,
     include_synonyms: bool = False,
     expand: bool = False,
@@ -2312,7 +2455,7 @@ def write_jsonld_context(
         json.dump(obj, file, indent=4, sort_keys=True)
 
 
-def _get_expanded_term(record: Record, *, expand: bool) -> Union[str, dict[str, Any]]:
+def _get_expanded_term(record: Record, *, expand: bool) -> str | dict[str, Any]:
     if not expand:
         return record.uri_prefix
 
@@ -2329,7 +2472,7 @@ def _get_expanded_term(record: Record, *, expand: bool) -> Union[str, dict[str, 
 
 def write_shacl(
     converter: Converter,
-    path: Union[str, Path],
+    path: str | Path,
     *,
     include_synonyms: bool = False,
 ) -> None:
@@ -2387,7 +2530,7 @@ def write_shacl(
 
 
 def write_tsv(
-    converter: Converter, path: Union[str, Path], *, header: tuple[str, str] = ("prefix", "base")
+    converter: Converter, path: str | Path, *, header: tuple[str, str] = ("prefix", "base")
 ) -> None:
     """Write a simple prefix map CSV file.
 
@@ -2424,7 +2567,7 @@ def write_tsv(
             writer.writerow((record.prefix, record.uri_prefix))
 
 
-def _get_shacl_line(prefix: str, uri_prefix: str, pattern: Optional[str] = None) -> str:
+def _get_shacl_line(prefix: str, uri_prefix: str, pattern: str | None = None) -> str:
     line = f'    [ sh:prefix "{prefix}" ; sh:namespace "{uri_prefix}"^^xsd:anyURI '
     if pattern:
         pattern = pattern.replace("\\", "\\\\")
@@ -2512,3 +2655,13 @@ def upgrade_prefix_map(prefix_map: Mapping[str, str]) -> list[Record]:
         Record(prefix=prefix, prefix_synonyms=prefix_synonyms, uri_prefix=uri_prefix)
         for uri_prefix, (prefix, *prefix_synonyms) in sorted(priority_prefix_map.items())
     ]
+
+
+def _get_converter_from_context(info: core_schema.ValidationInfo) -> Converter | None:
+    context = info.context or {}
+    if isinstance(context, Converter):
+        return context
+    elif isinstance(context, dict):
+        return context.get("converter")
+    else:
+        raise TypeError
